app/backend/pyproject.toml
[tool.poetry]
name = "Digital-desire"
version = "0.1.0"
description = ""
authors = ["Digital-desire"]

[tool.poetry.dependencies]
python = "3.8.10"
django = "3.*"
djangorestframework = "3.*"
black = "*"

[tool.poetry.dev-dependencies]

